request login ok
